Update libtomcrypt to https://github.com/libtom/libtomcrypt/commit/de15a6fad94a1d6d05a197bd409adbec084fac2b.
This is mainly with focus on issues with the ECC code. One of these issues is a "possible overflow in ecc_ansi_x963_export" (https://github.com/libtom/libtomcrypt/issues/58). Other than that all changes are minimal, only one new compile flag had to be added to get around a compiler incompatibility case for LTC_FAST (we're not working with GNU compilers here) and the code will fully compile.