[Maps] Don't wait for the entire database, just glass data
This actually allows the thread to continue and this 'fixes' the
material crash for custom maps. The Material_Sort function requires
something our thread seems to compute. I haven't had a deep look into it,
as for now this fix works, but for the future, a real fix is needed!