Discard non-existent dvar
The dvar `aim_lockon_pitch_strength` does not exist. Whether it should exist is something to consider, and if so, it should then be registered appropriately.

This commit removes all references to it and uses `aim_lockon_strength` as the scale/strength factor when calculating the pitch turn rate. This aligns with the behavior in the Xbox 360 version of the game.